// backend/src/models/point.model.ts
export interface Point {
    id?: number;
    deviceId?: number;
    name: string;
    address?: string;
    type?: string;
    dataType?: string;
    description?: string;
    units?: string;
    mapName?: string;
    sourceFile?: string;
    destination?: string;
    source?: string;
    pointType?: string;
    index?: string;
    comment?: string;
    createdAt?: Date;
}